--- src/main/java/com/mallang/mallang_backend/domain/sentence/expression/service/ExpressionService.java
@@ -1,9 +1,0 @@
-package com.mallang.mallang_backend.domain.sentence.expression.service;
-
-import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionResponse;
-
-import java.util.List;
-
-public interface ExpressionService {
-    List<ExpressionResponse> searchExpressions(String keyword);
-}
