+++ src/main/java/com/mallang/mallang_backend/domain/sentence/expressionbook/repository/ExpressionBookRepository.java
@@ -1,11 +1,12 @@
 package com.mallang.mallang_backend.domain.sentence.expressionbook.repository;
 
-import com.mallang.mallang_backend.domain.member.entity.Member;
-import com.mallang.mallang_backend.domain.sentence.expressionbook.entity.ExpressionBook;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.List;
 import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.mallang.mallang_backend.domain.member.entity.Member;
+import com.mallang.mallang_backend.domain.sentence.expressionbook.entity.ExpressionBook;
 
 public interface ExpressionBookRepository extends JpaRepository<ExpressionBook, Long> {
     List<ExpressionBook> findAllByMember(Member member);
