+++ src/main/java/com/mallang/mallang_backend/domain/sentence/expressionbook/service/ExpressionBookService.java
@@ -1,8 +1,10 @@
 package com.mallang.mallang_backend.domain.sentence.expressionbook.service;
 
+import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.DeleteExpressionsRequest;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionBookRequest;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionBookResponse;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionResponse;
+import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.MoveExpressionsRequest;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.savedExpressionsRequest;
 import jakarta.validation.Valid;
 
@@ -21,4 +23,9 @@ public interface ExpressionBookService {
   
     void save(savedExpressionsRequest request, Long expressionbookId);
 
+    void deleteExpressionsFromExpressionBook(DeleteExpressionsRequest request, Long memberId);
+
+    void moveExpressions(MoveExpressionsRequest request, Long memberId);
+
+    List<ExpressionResponse> searchExpressions(Long memberId, String keyword);
 }
